Move service interfaces to Services/Interfaces
Refactor service interfaces into a dedicated Services/Interfaces folder and namespace. Updated IDateTimeProvider, IHtmlSanitizerService and ISteamService paths/namespaces and adjusted all implementations and consumers (DateTimeProvider, HtmlSanitizerService, SteamService, SteamController) to import RandomSteamGame.Services.Interfaces. Also removed a stale Features folder entry from the .csproj. Improves project organization and clarifies interface locations.
